Moved arrayToObject and domainMatch to RPCUtil.
